+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), following [Se
 
 ## [Unreleased]
 
+### Changed
+
+- Add sanitize_args so credentials are not logged when running commands.
+
 ## [0.1.0] 2021-06-02
 
 ### added